EPBDS-13226 Remove XStream library due vulnerabilities and unsafe reflection operations
