+++ STUDIO/org.openl.rules.webstudio/src/org/openl/rules/project/model/v5_11/converter/ProjectDescriptorVersionConverter_v5_11.java
@@ -12,7 +12,7 @@ import org.openl.util.CollectionUtils;
 /**
  * @author nsamatov.
  */
-public class ProjectDescriptor_v5_11VersionConverter implements ObjectVersionConverter<ProjectDescriptor, ProjectDescriptor_v5_11> {
+public class ProjectDescriptorVersionConverter_v5_11 implements ObjectVersionConverter<ProjectDescriptor, ProjectDescriptor_v5_11> {
     private final ModuleVersionConverter moduleVersionConverter = new ModuleVersionConverter();
 
     @Override
