+++ STUDIO/org.openl.rules.webstudio/src/org/openl/rules/project/model/v5_12/converter/ProjectDescriptorVersionConverter_v5_12.java
@@ -1,4 +1,4 @@
-package org.openl.rules.project.model.v5_16.converter;
+package org.openl.rules.project.model.v5_12.converter;
 
 import java.util.List;
 
@@ -6,21 +6,21 @@ import org.openl.rules.project.model.Module;
 import org.openl.rules.project.model.ObjectVersionConverter;
 import org.openl.rules.project.model.ProjectDependencyDescriptor;
 import org.openl.rules.project.model.ProjectDescriptor;
+import org.openl.rules.project.model.v5_11.Module_v5_11;
+import org.openl.rules.project.model.v5_11.converter.ModuleVersionConverter;
 import org.openl.rules.project.model.v5_12.ProjectDependencyDescriptor_v5_12;
-import org.openl.rules.project.model.v5_12.converter.ProjectDependencyDescriptorVersionConverter;
-import org.openl.rules.project.model.v5_16.Module_v5_16;
-import org.openl.rules.project.model.v5_16.ProjectDescriptor_v5_16;
+import org.openl.rules.project.model.v5_12.ProjectDescriptor_v5_12;
 import org.openl.util.CollectionUtils;
 
 /**
  * @author nsamatov.
  */
-public class ProjectDescriptor_5_16VersionConverter implements ObjectVersionConverter<ProjectDescriptor, ProjectDescriptor_v5_16> {
+public class ProjectDescriptorVersionConverter_v5_12 implements ObjectVersionConverter<ProjectDescriptor, ProjectDescriptor_v5_12> {
     private final ModuleVersionConverter moduleVersionConverter = new ModuleVersionConverter();
     private final ProjectDependencyDescriptorVersionConverter dependencyConverter = new ProjectDependencyDescriptorVersionConverter();
 
     @Override
-    public ProjectDescriptor fromOldVersion(ProjectDescriptor_v5_16 oldVersion) {
+    public ProjectDescriptor fromOldVersion(ProjectDescriptor_v5_12 oldVersion) {
         ProjectDescriptor descriptor = new ProjectDescriptor();
         descriptor.setName(oldVersion.getName());
         descriptor.setComment(oldVersion.getComment());
@@ -30,7 +30,7 @@ public class ProjectDescriptor_5_16VersionConverter implements ObjectVersionConv
         descriptor.setModules(modules);
 
         List<ProjectDependencyDescriptor> dependencies = CollectionUtils.map(oldVersion.getDependencies(),
-            dependencyConverter::fromOldVersion);
+                dependencyConverter::fromOldVersion);
         if (CollectionUtils.isNotEmpty(dependencies)) {
             descriptor.setDependencies(dependencies);
         }
@@ -42,14 +42,14 @@ public class ProjectDescriptor_5_16VersionConverter implements ObjectVersionConv
     }
 
     @Override
-    public ProjectDescriptor_v5_16 toOldVersion(ProjectDescriptor currentVersion) {
-        ProjectDescriptor_v5_16 descriptor = new ProjectDescriptor_v5_16();
+    public ProjectDescriptor_v5_12 toOldVersion(ProjectDescriptor currentVersion) {
+        ProjectDescriptor_v5_12 descriptor = new ProjectDescriptor_v5_12();
 
         descriptor.setName(currentVersion.getName());
         descriptor.setComment(currentVersion.getComment());
         descriptor.setClasspath(currentVersion.getClasspath());
 
-        List<Module_v5_16> modules = CollectionUtils.map(currentVersion.getModules(),
+        List<Module_v5_11> modules = CollectionUtils.map(currentVersion.getModules(),
                 moduleVersionConverter::toOldVersion);
         descriptor.setModules(modules);
