+++ STUDIO/org.openl.rules.webstudio/src/org/openl/rules/project/xml/v5_16/PublisherTypeXmlAdapter_v5_16.java
@@ -1,0 +1,17 @@
+package org.openl.rules.project.xml.v5_16;
+
+import org.openl.rules.project.model.v5_16.RulesDeploy_v5_16;
+
+import javax.xml.bind.annotation.adapters.XmlAdapter;
+
+public class PublisherTypeXmlAdapter_v5_16 extends XmlAdapter<String, RulesDeploy_v5_16.PublisherType> {
+    @Override
+    public RulesDeploy_v5_16.PublisherType unmarshal(String name) {
+        return RulesDeploy_v5_16.PublisherType.valueOf(name.toUpperCase());
+    }
+
+    @Override
+    public String marshal(RulesDeploy_v5_16.PublisherType publisherType) {
+        return publisherType.toString();
+    }
+}
