+++ STUDIO/org.openl.rules.webstudio/test/org/openl/rules/project/xml/v5_11/XmlProjectDescriptorSerializer_v5_11Test.java
@@ -10,7 +10,7 @@ import java.util.stream.Collectors;
 import org.openl.rules.project.model.PathEntry;
 import org.openl.rules.project.model.ProjectDescriptor;
 import org.openl.rules.project.model.v5_11.ProjectDescriptor_v5_11;
-import org.openl.rules.project.model.v5_11.converter.ProjectDescriptor_v5_11VersionConverter;
+import org.openl.rules.project.model.v5_11.converter.ProjectDescriptorVersionConverter_v5_11;
 import org.openl.rules.project.xml.BaseProjectDescriptorSerializer;
 import org.openl.rules.project.xml.BaseProjectDescriptorSerializerTest;
 
@@ -25,7 +25,7 @@ public class XmlProjectDescriptorSerializer_v5_11Test {
         ProjectDescriptor projectDescriptor = BaseProjectDescriptorSerializerTest.initProjectDescriptorForTest();
 
         BaseProjectDescriptorSerializer<ProjectDescriptor_v5_11> descriptorSerializer_v5_11 =
-                new BaseProjectDescriptorSerializer<>(new ProjectDescriptor_v5_11VersionConverter(), ProjectDescriptor_v5_11.class);
+                new BaseProjectDescriptorSerializer<>(new ProjectDescriptorVersionConverter_v5_11(), ProjectDescriptor_v5_11.class);
         String xml = descriptorSerializer_v5_11.serialize(projectDescriptor);
         assertEquals(xml, PROJECT_DESCRIPTOR_V5_11_XML);
     }
@@ -72,7 +72,7 @@ public class XmlProjectDescriptorSerializer_v5_11Test {
     public void testDeserialize() throws Exception {
         BaseProjectDescriptorSerializer<ProjectDescriptor_v5_11> descriptorSerializer_v5_11 =
                 new BaseProjectDescriptorSerializer<>(
-                        new ProjectDescriptor_v5_11VersionConverter(), ProjectDescriptor_v5_11.class);
+                        new ProjectDescriptorVersionConverter_v5_11(), ProjectDescriptor_v5_11.class);
         ProjectDescriptor deserializedProject = descriptorSerializer_v5_11.deserialize(
                 new FileInputStream("test-resources/org.openl.rules.project.xml/project-descriptor_v5_11.xml"));
         ProjectDescriptor project = BaseProjectDescriptorSerializerTest.initProjectDescriptorForTest();
