+++ STUDIO/org.openl.rules.webstudio/test/org/openl/rules/project/xml/v5_11/XmlRulesDeploySerializer_v5_11Test.java
@@ -2,6 +2,9 @@ package org.openl.rules.project.xml.v5_11;
 
 import org.junit.Test;
 import org.openl.rules.project.model.RulesDeploy;
+import org.openl.rules.project.model.v5_11.RulesDeploy_v5_11;
+import org.openl.rules.project.model.v5_11.converter.RulesDeployVersionConverter_v5_11;
+import org.openl.rules.project.xml.BaseRulesDeploySerializer;
 
 import java.io.FileInputStream;
 import java.util.Map;
@@ -16,7 +19,8 @@ public class XmlRulesDeploySerializer_v5_11Test {
     @Test
     public void testReadRulesDeploy() throws Exception {
         FileInputStream fis = new FileInputStream("test-resources/org.openl.rules.project.xml/rules-deploy.xml");
-        XmlRulesDescriptorSerializer_v5_11 serializer = new XmlRulesDescriptorSerializer_v5_11();
+        BaseRulesDeploySerializer<RulesDeploy_v5_11> serializer =
+                new BaseRulesDeploySerializer<>(new RulesDeployVersionConverter_v5_11(), RulesDeploy_v5_11.class);
         RulesDeploy rulesDeploy = serializer.deserialize(fis);
 
         assertNotNull(rulesDeploy);
@@ -41,7 +45,8 @@ public class XmlRulesDeploySerializer_v5_11Test {
     @Test
     public void testWriteRulesDeploy() throws Exception {
         RulesDeploy rulesDeploy = generateRulesDeployForTest();
-        XmlRulesDescriptorSerializer_v5_11 serializer = new XmlRulesDescriptorSerializer_v5_11();
+        BaseRulesDeploySerializer<RulesDeploy_v5_11> serializer =
+                new BaseRulesDeploySerializer<>(new RulesDeployVersionConverter_v5_11(), RulesDeploy_v5_11.class);
         String value = serializer.serialize(rulesDeploy);
         assertEquals(EXPECTED_VALUE, value);
     }
