+++ STUDIO/org.openl.rules.webstudio/test/org/openl/rules/project/xml/v5_12/XmlProjectDescriptorSerializer_v5_12Test.java
@@ -6,7 +6,7 @@ import org.openl.rules.project.model.PathEntry;
 import org.openl.rules.project.model.ProjectDependencyDescriptor;
 import org.openl.rules.project.model.ProjectDescriptor;
 import org.openl.rules.project.model.v5_12.ProjectDescriptor_v5_12;
-import org.openl.rules.project.model.v5_12.converter.ProjectDescriptor_v5_12VersionConverter;
+import org.openl.rules.project.model.v5_12.converter.ProjectDescriptorVersionConverter_v5_12;
 import org.openl.rules.project.xml.BaseProjectDescriptorSerializer;
 import org.openl.rules.project.xml.BaseProjectDescriptorSerializerTest;
 
@@ -25,7 +25,7 @@ public class XmlProjectDescriptorSerializer_v5_12Test {
         ProjectDescriptor projectDescriptor = BaseProjectDescriptorSerializerTest.initProjectDescriptorForTest();
 
         String xml = new BaseProjectDescriptorSerializer<>(
-                new ProjectDescriptor_v5_12VersionConverter(), ProjectDescriptor_v5_12.class).serialize(projectDescriptor);
+                new ProjectDescriptorVersionConverter_v5_12(), ProjectDescriptor_v5_12.class).serialize(projectDescriptor);
         assertEquals(xml, PROJECT_DESCRIPTOR_V5_12_XML);
     }
 
@@ -77,7 +77,7 @@ public class XmlProjectDescriptorSerializer_v5_12Test {
     @Test
     public void testDeserialize() throws Exception {
         ProjectDescriptor deserializedProject = new BaseProjectDescriptorSerializer<>(
-                new ProjectDescriptor_v5_12VersionConverter(), ProjectDescriptor_v5_12.class).deserialize(
+                new ProjectDescriptorVersionConverter_v5_12(), ProjectDescriptor_v5_12.class).deserialize(
                 new FileInputStream("test-resources/org.openl.rules.project.xml/project-descriptor_v5_12.xml"));
         assertProjectDescriptorDeserializedCorrectly(deserializedProject);
     }
