+++ STUDIO/org.openl.rules.webstudio/test/org/openl/rules/project/xml/v5_15/XmlRulesDeploySerializer_v5_15Test.java
@@ -3,6 +3,9 @@ package org.openl.rules.project.xml.v5_15;
 import org.junit.Test;
 import org.openl.rules.project.model.RulesDeploy;
 import org.openl.rules.project.model.RulesDeploy.PublisherType;
+import org.openl.rules.project.model.v5_15.RulesDeploy_v5_15;
+import org.openl.rules.project.model.v5_15.converter.RulesDeployVersionConverter_v5_15;
+import org.openl.rules.project.xml.BaseRulesDeploySerializer;
 
 import java.io.FileInputStream;
 import java.util.Map;
@@ -17,7 +20,8 @@ public class XmlRulesDeploySerializer_v5_15Test {
     @Test
     public void testReadRulesDeploy() throws Exception {
         FileInputStream fis = new FileInputStream("test-resources/org.openl.rules.project.xml/rules-deploy.xml");
-        XmlRulesDescriptorSerializer_v5_15 serializer = new XmlRulesDescriptorSerializer_v5_15();
+        BaseRulesDeploySerializer<RulesDeploy_v5_15> serializer =
+                new BaseRulesDeploySerializer<>(new RulesDeployVersionConverter_v5_15(), RulesDeploy_v5_15.class);
         RulesDeploy rulesDeploy = serializer.deserialize(fis);
 
         assertNotNull(rulesDeploy);
@@ -47,7 +51,8 @@ public class XmlRulesDeploySerializer_v5_15Test {
     @Test
     public void testWriteRulesDeploy() throws Exception {
         RulesDeploy rulesDeploy = generateRulesDeployForTest();
-        XmlRulesDescriptorSerializer_v5_15 serializer = new XmlRulesDescriptorSerializer_v5_15();
+        BaseRulesDeploySerializer<RulesDeploy_v5_15> serializer =
+                new BaseRulesDeploySerializer<>(new RulesDeployVersionConverter_v5_15(), RulesDeploy_v5_15.class);
         String value = serializer.serialize(rulesDeploy);
         assertEquals(EXPECTED_VALUE, value);
     }
