+++ STUDIO/org.openl.rules.webstudio/test/org/openl/rules/project/xml/v5_23/XmlRulesDeploySerializer_v5_23Test.java
@@ -3,6 +3,9 @@ package org.openl.rules.project.xml.v5_23;
 import org.junit.Test;
 import org.openl.rules.project.model.RulesDeploy;
 import org.openl.rules.project.model.RulesDeploy.PublisherType;
+import org.openl.rules.project.model.v5_23.RulesDeploy_v5_23;
+import org.openl.rules.project.model.v5_23.converter.RulesDeployVersionConverter_v5_23;
+import org.openl.rules.project.xml.BaseRulesDeploySerializer;
 
 import java.io.FileInputStream;
 import java.util.Map;
@@ -17,7 +20,8 @@ public class XmlRulesDeploySerializer_v5_23Test {
     @Test
     public void testReadRulesDeploy() throws Exception {
         FileInputStream fis = new FileInputStream("test-resources/org.openl.rules.project.xml/rules-deploy.xml");
-        XmlRulesDescriptorSerializer_v5_23 serializer = new XmlRulesDescriptorSerializer_v5_23();
+        BaseRulesDeploySerializer<RulesDeploy_v5_23> serializer =
+                new BaseRulesDeploySerializer<>(new RulesDeployVersionConverter_v5_23(), RulesDeploy_v5_23.class);
         RulesDeploy rulesDeploy = serializer.deserialize(fis);
 
         assertNotNull(rulesDeploy);
@@ -48,7 +52,8 @@ public class XmlRulesDeploySerializer_v5_23Test {
     public void testWriteRulesDeploy() throws Exception {
         RulesDeploy rulesDeploy = generateRulesDeployForTest();
         rulesDeploy.setPublishers(new PublisherType[]{PublisherType.RMI, PublisherType.KAFKA});
-        XmlRulesDescriptorSerializer_v5_23 serializer = new XmlRulesDescriptorSerializer_v5_23();
+        BaseRulesDeploySerializer<RulesDeploy_v5_23> serializer =
+                new BaseRulesDeploySerializer<>(new RulesDeployVersionConverter_v5_23(), RulesDeploy_v5_23.class);
         String value = serializer.serialize(rulesDeploy);
         assertEquals(EXPECTED_VALUE, value);
     }
